requirements.txt
django-environ==0.11.2
djangorestframework-simplejwt==5.3.1